attempting to consolidate backbone file structure. Have added back in grabing players names.
