+++ src/app/collections/game_collection.js
@@ -1,1 +1,1 @@
- 
+
